+++ Boa.Constrictor.Playwright/CHANGELOG.md
@@ -16,3 +16,5 @@ Its format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Added `Click`, `Fill`, and `Go` Tasks
 - Added `OpenNewPage` task to initialize page and navigate to the specified url
 - Added `Text` and `Attribute` questions
+- Added an `Expects` extension method to `IActor` to make it easier to use playwrights built in `ILocatorAssertions`
+  - e.g. `await MyActor.Expects(WikiPage.Title).ToHaveTextAsync("Giant panda");`